[FUA-98] Move `UploadType` type declarations to account for circular dependency issues
