+++ src/renderer/state/types.ts
@@ -32,8 +32,7 @@ import {
   WellResponse,
 } from "../services/metadata-management-service/types";
 import { LocalStorage } from "../types";
-
-import { UploadType } from "./selection/types";
+import type { UploadType } from "../types";
 
 import Process = CreateLogic.Config.Process;
 import DepObj = CreateLogic.Config.DepObj;
