+++ src/renderer/state/upload/test/selectors.test.ts
@@ -6,7 +6,7 @@ import { AnnotationName } from "../../../constants";
 import { TemplateAnnotation } from "../../../services/metadata-management-service/types";
 import { UploadRequest } from "../../../services/types";
 import { Duration } from "../../../types";
-import { UploadType } from "../../selection/types";
+import { UploadType } from "../../../types";
 import {
   getMockStateWithHistory,
   mockAnnotationTypes,
